pyproject.toml
[project]
name = "ouraflask"
version = "0.0.1"
description = "A Flask server for the storage of OURA API data."
dependencies = ["flask", "python-dotenv", "requests", "requests-ouathlib"]

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"
